Basic loading of chart from repository

+++ Code/JsonGoals/GoalSpike/IWeightedEntity.cs
@@ -6,7 +6,7 @@ namespace JsonDb
 {
     public interface IWeightedEntity
     {
-        int Weighting { get; }
+        double Weighting { get; }
         double PercentCompleted();
     }
 }
